+++ src/input/sensors/sensor_library/arduino_rain_sensor/arduino_rain_sensor.h
@@ -6,10 +6,10 @@
 
 /* Define the digital output value indicating rain detected by the sensor, used in digital read mode */
 /* NOTE: The sensor uses reverse logic — 0 means rain is detected, and 1 means no rain */
-#define ARDUINO_RAIN_SENSOR_RAIN_DETECTED    (0)
+#define ARDUINO_RAIN_SENSOR_RAIN_DETECTED    (int)(0)
 
 /* Define the threshold value for analog readings; values below this indicate rain */
-#define ARDUINO_RAIN_SENSOR_ANALOG_THRESHOLD (500u)
+#define ARDUINO_RAIN_SENSOR_ANALOG_THRESHOLD (int)(500u)
 
 /**
  * @brief Initializes the rain sensor.
