+++ src/output/serial_console/serial_console_config.h
@@ -7,6 +7,6 @@
  * Defines the baud rate for serial communication - UART or USB cable.
  * Set the same baud rate on your device or change it here to accomodate it.
  */
-#define SERIAL_CONSOLE_BAUDRATE (9600u)
+#define SERIAL_CONSOLE_BAUDRATE (uint64_t)(9600u)
 
 #endif
